+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Head
+
+* Added: support for extensions on `.stylelintrc` files (by upgrading cosmiconfig).
+
 # 3.2.3
 
 * Fixed: `selector-no-vendor-prefix` now handles custom-property-sets.